clean up collaborators
